Separate out futures & streams.
